requirements.txt
Django==1.11.1
markdown==2.6.8
django-markdown-app==0.9.2
djangorestframework==3.6.3
django-qiniu-storage==2.3.0
mysql-python==1.2.5
Pillow==4.0.0
django-simple-captcha==0.5.3
django-extensions==1.7.6
bootstrap-admin==0.3.7.1
pylibmc==1.5.1
django-ratelimit==1.0.1
raven==6.0.0
django-filter==1.0.4